Fix RadioButtonsOptions builder parameter type. Add race lockout to new and apply buttons.

+++ soh/soh/SohGui/UIWidgets.hpp
@@ -414,7 +414,7 @@ struct RadioButtonsOptions : WidgetOptions {
         color = color_;
         return *this;
     }
-    RadioButtonsOptions& DefaultIndex(float defaultIndex_) {
+    RadioButtonsOptions& DefaultIndex(int32_t defaultIndex_) {
         defaultIndex = defaultIndex_;
         return *this;
     }
